backend/requirements.txt
langchain==0.3.27
langchain_core==0.3.77
langgraph==0.6.8
openai==2.0.1
pydantic==2.11.9
python-dotenv==1.1.1
typing_extensions==4.15.0
langgraph>=0.1.0
langgraph-sdk>=0.1.0
